+++ modules/runtime/src_static/binary/build.reader.cpp
@@ -280,7 +280,7 @@ int ReadTrait<skr::string_view>::Read(skr_binary_reader_t* reader, skr_blob_aren
     if (ret != 0)
         return ret;
     auto pstr = (const char8_t*)arena.get_buffer() + offset;
-    str = skr::string_view(pstr, (uint64_t)size);
+    str = skr::string_view(pstr, (int32_t)size);
     // null terminate
     auto ptr = const_cast<char8_t*>(str.u8_str() + size);
     *ptr = u8'\0';
